+++ growth-value-calculator/src/main/java/the/best/growthvaluecalculator/controller/CalculatorController.java
@@ -1,7 +1,6 @@
 package the.best.growthvaluecalculator.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -10,7 +9,6 @@ import the.best.growthvaluecalculator.responses.CalculatorResponse;
 import the.best.growthvaluecalculator.util.Calculator;
 
 @RestController//8000
-//@CrossOrigin(origins = "*")
 public class CalculatorController {
 	
 	@GetMapping("/stock-fcf")
@@ -114,7 +112,7 @@ public class CalculatorController {
 	public double calculateFixedValue(
 			@RequestParam long principle, 
 			@RequestParam double interest,
-			@RequestParam int length,
+			@RequestParam double length,
 			@RequestParam int compoundFrequency
 			) {
 		return Calculator.calculateTotal(principle, interest, length, compoundFrequency);
